+++ resources/gaming-fe/src/App.tsx
@@ -1,268 +1,15 @@
-import { useEffect, useState } from 'react';
-
 import Game from './components/Game';
-import WalletConnectHeading from './components/WalletConnectHeading';
-import { blockchainDataEmitter } from './hooks/BlockchainInfo';
-import { BlockchainReport } from './types/ChiaGaming';
-import { getSaveList, loadSave } from './hooks/save';
-import {
-  getGameSelection,
-  getSearchParams,
-  generateOrRetrieveUniqueId,
-} from './util';
-import GameRedirectPopup from './components/GameRedirectPopup';
-import { Loader2 } from 'lucide-react';
+import ParentFrame from './components/ParentFrame';
+import { getSearchParams } from './util';
 
 const App = () => {
-  const uniqueId = generateOrRetrieveUniqueId();
-  const gameSelection = getGameSelection();
   const params = getSearchParams();
-  let useParams = params;
-  let useIframeUrl = 'about:blank';
-  // const saveList = getSaveList();
-  const saveList: string[] = []; // Disable save / reload
-  const shouldRedirectToLobby =
-    saveList.length == 0 && !params.lobby && !params.iStarted;
-  if (saveList.length > 0) {
-    const decodedSave = loadSave(saveList[0]);
-    if (decodedSave) {
-      useParams = decodedSave.searchParams;
-      useIframeUrl = decodedSave.url;
-    }
-  }
-  const [havePeak, setHavePeak] = useState(false);
-  const [iframeUrl, setIframeUrl] = useState(useIframeUrl);
-  const [fetchedUrls, setFetchedUrls] = useState(false);
-  const [iframeAllowed, setIframeAllowed] = useState('');
-  const [showPopup, setShowPopup] = useState(false);
-  const [pendingGameUrl, setPendingGameUrl] = useState<string | null>(null);
 
-  useEffect(() => {
-    const subscription = blockchainDataEmitter.getObservable().subscribe({
-      next: (_peak: BlockchainReport) => {
-        setHavePeak(true);
-      },
-    });
-
-    return () => subscription.unsubscribe();
-  });
-
-  // Fetch the urls document and get the tracker url so we know to allow the iframe
-  // to use the clipboard.
-  useEffect(() => {
-    if (!fetchedUrls) {
-      setFetchedUrls(true);
-      fetch('/urls')
-        .then((res) => res.json())
-        .then((urls) => {
-          let trackerURL = new URL(urls.tracker);
-          setIframeAllowed(trackerURL.origin);
-        })
-        .catch(e => console.error('[App] failed to fetch /urls for iframe origin:', e));
-    }
-  }, [fetchedUrls]);
-
-  // Redirect to the lobby if we haven't been given enough information to render
-  // the game yet.
-  //
-  // This will be inside a frame whose parent owns the wallet and blockchain
-  // connection soon.  I think we can change the iframe location from the outside
-  // in that scenario.
-  useEffect(() => {
-    if (shouldRedirectToLobby) {
-      fetch('/urls')
-        .then((res) => res.json())
-        .then((urls) => {
-          const baseUrl = urls.tracker;
-          const gameUrl = gameSelection
-            ? `${baseUrl}&uniqueId=${uniqueId}&token=${gameSelection.token}&view=game`
-            : `${baseUrl}&view=game&uniqueId=${uniqueId}`;
-
-          if (params.join) {
-            // It's an invite → wait for user to accept
-            setPendingGameUrl(gameUrl);
-            setShowPopup(true);
-          } else {
-            // No invite → go straight to game
-            setIframeUrl(gameUrl);
-          }
-        })
-        .catch(e => console.error('[App] failed to fetch /urls for lobby redirect:', e));
-    }
-  }, []);
-  // no dependency on params so URL stays the same
-
-  // Keep iframe in sync with the parent theme (CSS variables and dark class).
-  // If the iframe is same-origin we copy the CSS custom properties and `dark` class
-  // into the iframe's root. If cross-origin, we post a message the iframe can
-  // listen for and apply itself.
-  useEffect(() => {
-    let observer: MutationObserver | null = null;
-
-    function buildVarsPayload() {
-      const styles = getComputedStyle(document.documentElement);
-      const vars: Record<string, string> = {};
-      for (let i = 0; i < styles.length; i++) {
-        const name = styles[i];
-        if (name && name.startsWith('--')) {
-          vars[name] = styles.getPropertyValue(name).trim();
-        }
-      }
-      return vars;
-    }
-
-    async function syncThemeToIframe() {
-      const iframe = document.getElementById(
-        'subframe',
-      ) as HTMLIFrameElement | null;
-      if (!iframe) return;
-
-      const payload = buildVarsPayload();
-      const isDark = document.documentElement.classList.contains('dark');
-
-      try {
-        // Try same-origin access
-        const doc = iframe.contentDocument;
-        if (doc && doc.documentElement) {
-          const targetRoot = doc.documentElement;
-          // copy variables
-          Object.keys(payload).forEach((k) => {
-            targetRoot.style.setProperty(k, payload[k]);
-          });
-          // copy dark class
-          if (isDark) targetRoot.classList.add('dark');
-          else targetRoot.classList.remove('dark');
-          return;
-        }
-      } catch (e) {
-        // Access denied -> cross-origin, fall back to postMessage
-      }
-
-      // Cross-origin fallback: send theme data via postMessage
-      try {
-        iframe.contentWindow?.postMessage(
-          { type: 'theme-sync', vars: payload, dark: isDark },
-          '*',
-        );
-      } catch (e) {
-        // ignore
-      }
-    }
-
-    // Run on iframe load
-    function onLoad() {
-      syncThemeToIframe();
-    }
-
-    const iframeEl = document.getElementById(
-      'subframe',
-    ) as HTMLIFrameElement | null;
-    iframeEl?.addEventListener('load', onLoad);
-
-    // Respond to explicit requests from iframes that may have attached their
-    // message listener after the parent's initial postMessage. When an iframe
-    // posts {type: 'theme-request'}, send the current theme to it.
-    function messageHandler(ev: MessageEvent) {
-      try {
-        if (ev.data && ev.data.type === 'theme-request') {
-          // Only respond to requests coming from the iframe we care about.
-          // If there are multiple iframes, more checks may be needed.
-          syncThemeToIframe();
-        }
-      } catch (e) {
-        // ignore
-      }
-    }
-    window.addEventListener('message', messageHandler);
-
-    // Observe changes to the root class (for dark mode toggles) and resync
-    observer = new MutationObserver(() => {
-      syncThemeToIframe();
-    });
-    observer.observe(document.documentElement, {
-      attributes: true,
-      attributeFilter: ['class'],
-    });
-
-    // initial sync (in case iframe already loaded)
-    setTimeout(syncThemeToIframe, 150);
-
-    return () => {
-      iframeEl?.removeEventListener('load', onLoad);
-      window.removeEventListener('message', messageHandler);
-      if (observer) observer.disconnect();
-    };
-  }, [iframeUrl]);
-
-  const handleAccept = () => {
-    if (pendingGameUrl) {
-      setIframeUrl(pendingGameUrl);
-    }
-    setShowPopup(false);
-  };
-
-  const handleCancel = () => {
-    setShowPopup(false);
-    // Redirect to lobby
-    window.location.href = '/?lobby=1';
-  };
-
-  if (params.game && !params.join && !showPopup) {
+  if (params.game && !params.join) {
     return <Game params={params} />;
   }
 
-  const wcHeading = (
-    <div className='flex shrink-0 h-12 w-full'>
-      <WalletConnectHeading />
-    </div>
-  );
-
-  const pre_lobby_status = (
-    <div
-      /*className="flex flex-col relative w-screen h-screen"*/
-      className='w-full flex-1 border-0 m-0 p-0 flex flex-col items-center justify-center text-center gap-3'
-      style={{
-        backgroundColor: 'var(--color-canvas-bg-subtle)',
-        display: 'flex', // Enables flexbox
-        alignItems: 'center', // Centers children vertically
-        justifyContent: 'center', // Optional: centers children horizontally as well
-        height: '100vh', // Optional: ensures the container takes up the full viewport height
-      }}
-    >
-      <Loader2 className='h-6 w-6 z-0 animate-spin text-primary mb-4' />
-      Waiting for blockchain peak ...
-    </div>
-  );
-
-  if (!havePeak) {
-    return (
-      <div className='flex flex-col relative w-screen h-screen bg-canvas-bg-subtle'>
-        {wcHeading}
-        {pre_lobby_status}
-      </div>
-    );
-  }
-
-  return (
-    <div className='flex flex-col relative w-screen h-screen bg-canvas-bg-subtle'>
-      {wcHeading}
-      <div className='relative z-0 w-full flex-1 bg-canvas-bg-subtle'>
-        <iframe
-          id='subframe'
-          className='w-full h-full border-0 m-0 md:py-0 py-6 bg-canvas-bg-subtle'
-          src={iframeUrl}
-          allow={`clipboard-write self ${iframeAllowed}`}
-        ></iframe>
-      </div>
-      <GameRedirectPopup
-        open={showPopup}
-        gameName={params.game} // "calpoker"
-        message='You have been invited to join this game.'
-        onAccept={handleAccept}
-        onCancel={handleCancel}
-      />
-    </div>
-  );
+  return <ParentFrame params={params} />;
 };
 
 export default App;
